+ Fix a bug in sorting results of elasticsearch queries + Update a feature
In this commit the feature that prevented users from searching with queries of low character count (less or equal than 2) for both message queries and inline queries was updated.

A bug in `tase/db/elasticsearch_modeles/audio.py` module that caused sorting results of audios query to be incorrect, was also fixed.