+ Add feature + Reformat Codes + Refactor Codes
Added a feature to extract usernames from chats on a scheduled interval add them to a buffer to be checked and add to the database for indexing if it does not violate current indexing rules. Moreover, Added a package named `enums` in the `db` package to unify usage of enum types across different database clients. On top of that, refactored and updated some codes to for better readability.