+ Add feature + Refactor codes + Bug fix
Added a feature to index audio files that are not shown in the `audios` section of telegram, instead, they can be seen in the `files/documents` section. However, most of these audio files are not valid for inline mode due to lack of attributes like `title`, etc... . Furthermore, added a new feature to `BaseCommand` class to support running commands from a callback query function.