+ Add feature + Update codes + Fix bug
Added the feature to store more metrics in the `UsernameExtractorMetadata` class to calculate a more accurate score. In addition, I added a few more methods to improve the class functionality.
In `Mentions` class added a new field named `is_checked` to account for checking functionality.
In `graph_db` module updated the method which updated the username extractor metadata to check for conflicts in updates and try to resolves them if possible.